move the class over to its own file, autoload it with composer, PSR-4
