+++ aspen-book-list.php
@@ -17,6 +17,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
+
+// LOAD the AUTOLOADING class
+require_once dirname(__FILE__)."/vendor/autoload.php";
+
+
 /**
  * Registers the block using the metadata loaded from the `block.json` file.
  * Behind the scenes, it registers also all assets so they can be enqueued
