+++ composer.json
@@ -1,0 +1,16 @@
+{
+    "name": "cpl/aspen-book-list",
+    "type": "project",
+    "autoload": {
+        "psr-4": {
+            "Cpl\\AspenBookList\\": "includes/"
+        }
+    },
+    "authors": [
+        {
+            "name": "Will Skora",
+            "email": "175735853+skora-cpl@users.noreply.github.com"
+        }
+    ],
+    "require": {}
+}
