+++ src/aspen-book-list/render.php
@@ -24,71 +24,7 @@
 
 $list_id = isset( $attributes['listID'] ) ? $attributes['listID'] : '';
 
-final class AspenList {
-
-	/**
-	 * Get the aspen CATALOG URL, return empty string if not defined
-	 *
-	 * @return string aspen catalog url
-	 */
-	protected function get_aspen_url(): string {
-		return defined( 'ASPEN_API_CATALOG_URL' ) ? constant( 'ASPEN_API_CATALOG_URL' ) : '';
-	}
-
-	/**
-	 * Get the associated/ special token that you registered with your Aspen Support Team; return empty string if not defined
-	 *
-	 * @return string special token
-	 */
-	protected function get_aspen_special_token(): string {
-		return defined( 'ASPEN_API_AUTHORIZATION_TOKEN' ) ? constant( 'ASPEN_API_AUTHORIZATION_TOKEN' ) : '';
-	}
-
-	// public function fetch_prep() {
-
-	// }
-
-	// returns an array, teh_request
-	/**
-	* @param  // listid attribute
-	* @return array|WP_Error
-	*/
-	public function fetch_list() {
-		$teh_request = wp_remote_get(
-		$this->get_aspen_url() . '/API/ListAPI?method=getListTitles&id=26',
-		[
-		'headers' => [
-			'X-Custom-SPECIAL' => $this->get_aspen_special_token(),
-			'X-custom-CPL'     => 'WP_PHP_Aspen_block',
-		],
-		]
-	);
-
-	// set transient for the list spefifically
-	// set_transient( self::LIBCAL_ACCESS_TOKEN_CACHE_GROUP, $data['access_token'], 3600 );
-
-
-	// Check for request errors
-	if ( is_wp_error( $teh_request ) ) {
-		return new \WP_Error( 'libcal_token_error', __( 'Failed to retrieve LibCal access token.', 'cpl-plugin' ) );
-	}
-
-	$body = wp_remote_retrieve_body( $teh_request );
-
-	$teh_data = json_decode( $body, true );
-
-	return $teh_data;
-
-	}
-
-}
-
-
-
-
-
-
-$my_new_list_object = new AspenList();
+$my_new_list_object = new \Cpl\AspenBookList\AspenList();
 $my_new_list_object->fetch_list();
 
 
@@ -125,7 +61,7 @@ settype( $list_id, 'integer' );
 // $teh_data = json_decode( $body, true );
 
 // even there is a successful response, the JSON body can return a message that has an error message
-if ( $teh_data['result']['success'] === false) {
+// if ( $teh_data['result']['success'] === false) {
 	// $the_error_message = $teh_data['result']['message'];
 	// $the_error_message = new \WP_Error;
 	// $the_error_message->add( $teh_data['result']['message'] , $teh_data['result']['message'] );
@@ -133,7 +69,7 @@ if ( $teh_data['result']['success'] === false) {
 
 		// throw new Exception ( $the_error_message . 'This happens when the list is set to private; OR You haven\'t entered the list\'s ID. please toggle the list to be public so it can be used');
 	// return new WP_Error( $the_error_message, 'Falling and cant get up' );
-}
+// }
 
 if ( ! empty( $teh_data ) ) {
 	?>
